fix (init/install): fix install fail with no cache
Fix incorrectly failing to init new projects if there is no existing
version and no version is passed in. In this case, info.tag was being
set to the latest known version, but the undefined version variable
was still being used to determine if any versions were available.

This now compares the existing version against the latest version
available in this case, then installs the latest version (instead of
skipping to the 'no valid versions found' failure case).

Additionally, this adds the module name to the 'no valid versions found'
failure message when a valid version really could not be found.