chore (simulator): stop icon not found errors
Catch file not found errors in simulator /icon requests.

Devkit falls back to a hardcoded resources/images/defaultIcon.png file
if no icon is specified in the manifest, which prints ENOENT errors when
serving devkit for every project that doesn't have an icon set and
doesn't have a matching default.

A more complete solution would be to refactor all the paths to the
default icon and add an additional fallback default in devkit if a
default is desired.