Add error propagation to InstallCommand
This is the apex of all of the promisification and additional error
handling throughout the devkit app / module management code. Commands
can now filter on any number of error types in order to display helpful
error messages to the user. By setting the environment variable
DEVKIT_TRACE, users can get extra information (such as a stack trace)
for debugging purposes.