Fix issue where all builds were debug
The change to the `src/devkit.js` hashbang to be debian-friendly caused
devkit to be invoked using the full path to the node.js executable
instead of just the binary name. The debug command, which apparently
also handles release builds, checked that the first argument had the
**exact value** of _node_. For now, it checks that the first argument
_contains_ node. This solution will handle both relative and absolute
paths of both `node` and `nodejs` binaries.

The command system should be updated so that lower level argument
parsing is handled elsewhere than in the command logic.

Resolves #33.