warn about untracked files
Previously we prevented git checkouts for modules with untracked and modified
files.  Untracked files are normally ok and may be the result of normal module
processes (though these files should be in .gitignore!).  Therefore, we now
only prevent git checkouts for modules with modified files.  We still log a
warning for untracked files in case they actually prevent a checkout (the
recommendation is to remove them and try again in case of a failure).