fix the detection for existing clients
The current detection is redundant and incorrect in its detection.
It seems to connect a client and then check if it exists.
This would always return true which means every client is rejected

The easiest way to manage the connected users atm is to just to
remove and add the UUID from the list